kubernetes/test/requirements.txt
Flask==2.1.2
Werkzeug==2.0.2
redis
prometheus_flask_exporter
mysql-connector-python